PIP-117 macierz zwraca dane na podstawie query

+++ src/main/java/net/amg/jira/plugins/velocity/Task.java
@@ -3,10 +3,10 @@ package net.amg.jira.plugins.velocity;
 public class Task {
 	private String name;
 	private String url;
-	private int probability;
-	private int consequency;
+	private Double probability;
+	private Double consequency;
 	
-	public Task(String name, String url, int probability, int consequency) {
+	public Task(String name, String url, Double probability, Double consequency) {
 		super();
 		this.name = name;
 		this.url = url;
@@ -26,16 +26,16 @@ public class Task {
 	public void setUrl(String url) {
 		this.url = url;
 	}
-	public int getProbability() {
+	public Double getProbability() {
 		return probability;
 	}
-	public void setProbability(int probability) {
+	public void setProbability(Double probability) {
 		this.probability = probability;
 	}
-	public int getConsequency() {
+	public Double getConsequency() {
 		return consequency;
 	}
-	public void setConsequency(int consequency) {
+	public void setConsequency(Double consequency) {
 		this.consequency = consequency;
 	}
